Add: execfork.c and getcmd.c, moved declarations for these files to header. Added more information to shell prompt.
