+++ shell.h
@@ -14,6 +14,8 @@
 void dsh_read_line(char **buf);
 void freestuff(char **args, size_t *wc, char *buf, char **paths, size_t *pathc);
 void strbrk(char *buf, char ***args, const char delim, size_t *wc);
+void getcmd(char **args, char **envp, char ***pathsplit, size_t *pathc);
+void execfork(char **envp, char **args);
 /* Subroutines */
 void countwords(char *buf, size_t *wc, const char dl, unsigned int *slen);
 void squeeze_spaces(char *str_d, unsigned int *slen);
